object-detection/requirements.txt
click==8.1.7
easyocr==1.7.1
matplotlib
numpy==1.24.3
omni==0.0.2
open3d==0.13.0
opencv_contrib_python==4.7.0.68
opencv_python==4.9.0.80
Pillow==10.2.0
pyrealsense2==2.54.2.5684
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0
setuptools==69.0.3
ultralytics==8.1.8
